# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zelon API",
    "version": "0.3.1",
    "description": "Self-hosted WhatsApp Web integration. Keys are scoped to one instance. Dashboard-only operations require a session and matching Origin. WhatsApp provider rules apply."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "instanceKey": {
        "type": "http",
        "scheme": "bearer"
      },
      "consoleSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zelon_session"
      }
    },
...
  },
  "paths": {
    "/api/login": {
      "post": {
        "operationId": "post__api_login",
        "summary": "POST /api/login",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "operationId": "get__api_me",
        "summary": "GET /api/me",
        "security": [
          {
            "consoleSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/account/password": {
      "put": {
        "operationId": "put__api_account_password",
        "summary": "PUT /api/account/password",
        "security": [
          {
            "consoleSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "operationId": "post__api_logout",
        "summary": "POST /api/logout",
        "security": [
          {
            "consoleSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "operationId": "get__api_admin_users",
        "summary": "GET /api/admin/users",
        "security": [
          {
            "consoleSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "post": {
        "operationId": "post__api_admin_users",
        "summary": "POST /api/admin/users",
        "security": [
          {
            "consoleSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/admin/system": {
      "get": {
        "operationId": "get__api_admin_system",
        "summary": "GET /api/admin/system",
        "security": [
          {
            "consoleSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/admin/users/{userId}": {
      "put": {
        "operationId": "put__api_admin_users__userId",
        "summary": "PUT /api/admin/users/:userId",
        "security": [
          {
            "consoleSession": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{userId}/reset-password": {
      "post": {
        "operationId": "post__api_admin_users__userId_reset_password",
        "summary": "POST /api/admin/users/:userId/reset-password",
        "security": [
          {
            "consoleSession": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/instances/{id}/restore": {
      "post": {
        "operationId": "post__api_instances__id_restore",
        "summary": "POST /api/instances/:id/restore",
        "security": [
          {
            "consoleSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/archived-instances": {
      "get": {
        "operationId": "get__api_archived_instances",
        "summary": "GET /api/archived-instances",
        "security": [
          {
            "consoleSession": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/instances/{id}/name": {
      "put": {
        "operationId": "put__api_instances__id_name",
        "summary": "PUT /api/instances/:id/name",
        "security": [
          {
            "consoleSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Operation requires dashboard/admin privileges"
          },
          "404": {
            "description": "Record not found or outside your instance"
          },
          "409": {
            "description": "Connection/state/idempotency conflict"
          },
          "429": {
            "description": "Rate limited"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  }
}